Add a chartview for an overview of activities and how time was spent
* Later add possibility to use results on monthly, weekly and yearly base and let user decide which month, week, year

+++ Hemera/Hemera/Models/Activity.cs
@@ -154,8 +154,8 @@ namespace Hemera.Models
 
         public TimeType DurationType;
 
-        private double _Duration = 60;
-        public double Duration
+        private float _Duration = 60;
+        public float Duration
         {
             get => _Duration;
             set
